Found by Jill Jones:
I think the problem is in lsb/src/tests/libchk/libchk.c.  The "if"
statements at lines 77 and 78 don't have any braces associated
with them, so the "continue" at line 82 will always be executed
every time through the loop, and the symbol versions check
later in the loop will never be executed.